backend/requirements.txt
flask
flask-cors
pymongo
bcrypt
PyJWT
python-dotenv
google-auth
gunicorn
